Added sales tax to the automatically created invoices. Resolves #66.

+++ NumberSearch.DataAccess/InvoiceNinja/Invoice.cs
@@ -210,7 +210,7 @@ namespace NumberSearch.DataAccess.InvoiceNinja
         public int updated_at { get; set; }
         public object archived_at { get; set; }
         public string notes { get; set; }
-        public float cost { get; set; }
+        public decimal cost { get; set; }
         public decimal qty { get; set; }
         public string tax_name1 { get; set; }
         public int tax_rate1 { get; set; }
